for sergey

+++ 4-module/2-task/server.js
@@ -7,8 +7,17 @@ const fs = require('fs');
 const LimitSizeStream = require('./LimitSizeStream');
 const server = new http.Server();
 
-server.on('request', (req, res) => {
+const emit = server.emit;
 
+server.on('request', (req, res) => {
+  req.emit = (...args) => {
+    console.log('req event', args[0]);
+    emit.apply(req, args);
+  };
+  res.emit = (...args) => {
+    console.log('res event', args[0]);
+    emit.apply(res, args);
+  };
   if (req.method !== 'POST') {
     res.statusCode = 500;
     res.end('Not implemented!');
@@ -31,6 +40,10 @@ server.on('request', (req, res) => {
   }
 
   const writeStream = fs.createWriteStream(filepath);
+  writeStream.emit = (...args) => {
+    console.log('writeStream event', args[0]);
+    emit.apply(writeStream, args);
+  };
   const limitStream = new LimitSizeStream({ limit: 1e6 });
 
   limitStream.on('error', (e) => {
@@ -52,16 +65,23 @@ server.on('request', (req, res) => {
 
   });
 
-  req.on('aborted', () => {
-    fs.unlinkSync(filepath);
-    return;
-  });
+  // req.on('aborted', () => {
+  //   fs.unlinkSync(filepath);
+  //   return;
+  // });
 
   // res.on('close', () => {
   //   if (!writeStream.writableFinished) {
   //     fs.unlinkSync(filepath);
   //   }
   // });
+
+  res.on('close', () => {
+    if (req.aborted) {
+      fs.unlinkSync(filepath);
+      return;
+    }
+  });
 
   writeStream.on('finish', () => {
     res.statusCode = 201;
